requirements.txt
pyngrok
numpy
pandas
scikit-learn
tensorflow
matplotlib
pyserial